add function support in where condition
