+++ src/main/java/org/nlpcn/es4sql/parse/ChildrenType.java
@@ -43,7 +43,7 @@ public class ChildrenType {
             this.simple = true;
         } else {
             Where where = Where.newInstance();
-            new SqlParser().parseWhere(secondParameter,where);
+            new WhereParser(new SqlParser()).parseWhere(secondParameter,where);
             if(where.getWheres().size() == 0)
                 throw new SqlParseException("unable to parse filter where.");
             this.where = where;
