+++ src/main/java/org/nlpcn/es4sql/parse/SelectParser.java
@@ -1,0 +1,7 @@
+package org.nlpcn.es4sql.parse;
+
+/**
+ * Created by allwefantasy on 9/2/16.
+ */
+public class SelectParser {
+}
